fix(vision): break circular import with vision_base module
Extract VisionProvider/VisionAnalysisResult/prompts to vision_base.py.
All providers import from vision_base, factory uses lazy imports.
