+++ src/scenefab/services/ai/providers/gemini35_flash.py
@@ -21,7 +21,7 @@ from typing import Any
 
 import httpx
 
-from ..vision_providers import VisionProvider, VisionAnalysisResult
+from ..vision_base import VisionProvider, VisionAnalysisResult
 
 logger = logging.getLogger(__name__)
