+++ src/scenefab/services/ai/providers/qwen37.py
@@ -18,7 +18,7 @@ from typing import Any
 
 from openai import OpenAI
 
-from ..vision_providers import VisionProvider, VisionAnalysisResult
+from ..vision_base import VisionProvider, VisionAnalysisResult
 
 logger = logging.getLogger(__name__)
